Rename small db.
